% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LM22}
\alias{LM22}
\title{Siganature matrix}
\format{A data frame with 547 rows and 22 variables:
\describe{
  \item{B.cells.naive}{naive B cells}
  \item{B.cells.memory}{memory B cells}
  \item{Plasma.cells}{Plasma cells}
  \item{T.cells.CD8}{CD8 T cells}
  \item{T.cells.CD4.naive}{naive CD4 T cells}
  \item{T.cells.CD4.memory.resting}{resting memory CD4 T cells}
  \item{T.cells.CD4.memory.activated}{activated memory CD4 T cells}
  \item{T.cells.follicular.helper}{follicular helper T cells }
  \item{T.cells.regulatory.Tregs.}{regulatory T cells}
  \item{T.cells.gamma.delta}{gamma delta T cells}
  \item{NK.cells.resting}{resting natural killer cells}
  \item{NK.cells.activated}{activated natural killer cells}
  \item{Monocytes}{monocytes}
  \item{Macrophages.M0}{M0 macrophages}
  \item{Macrophages.M1}{M1 macrophages}
  \item{Macrophages.M2}{M2 macrophages}
  \item{Dendritic.cells.resting}{resting dendritic cells}
  \item{Dendritic.cells.activated}{activated dendritic cells}
  \item{Mast.cells.resting}{resting mast cells}
  \item{Mast.cells.activated}{activated mast cells}
  \item{Eosinophils}{eosinophils}
  \item{Neutrophils}{neutrophils}
}}
\usage{
LM22
}
\description{
A dataset containing 547 genes and 22 TILs.
}
\references{
Aaron M. Newman, Chih Long Liu, Michael R. Green, Andrew J. Gentles, Weiguo Feng, Yue Xu, Chuong D. Hoang, Maximilian Diehn and Ash A. Alizadeh. Robust enumeration of cell subsets from tissue expression profiles.
}
\keyword{datasets}
