% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixture}
\alias{mixture}
\title{Gene-expression data from 14 follicular lymphoma patients}
\format{A data frame with 19416 rows and 14 variables:
\describe{
  \item{GSM1587831}{FL lymph node biopsy, untreated, 1063}
  \item{GSM1587832}{FL lymph node biopsy, untreated, 1080}
  \item{GSM1587833}{FL lymph node biopsy, untreated, 575}
  \item{GSM1587834}{FL lymph node biopsy, untreated, 581}
  \item{GSM1587835}{FL lymph node biopsy, untreated, 598}
  \item{GSM1587836}{FL lymph node biopsy, untreated, 639}
  \item{GSM1587837}{FL lymph node biopsy, untreated, 664}
  \item{GSM1587838}{FL lymph node biopsy, untreated, 666}
  \item{GSM1587839}{FL lymph node biopsy, untreated, 695}
  \item{GSM1587840}{FL lymph node biopsy, untreated, 706}
  \item{GSM1587841}{FL lymph node biopsy, untreated, 726}
  \item{GSM1587842}{FL lymph node biopsy, untreated, 731}
  \item{GSM1587843}{FL lymph node biopsy, untreated, 944}
  \item{GSM1587844}{FL lymph node biopsy, untreated, 959}
}}
\usage{
mixture
}
\description{
This gene-expression dataset consists of surgical lymph node biopsies of 14 
follicular lymphoma patients with 19416 genes. It is available on Gene 
Expression Omnibus (GEO) with accession number GSE65135.
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE65135}.
}
\references{
Aaron M. Newman, Chih Long Liu, Michael R. Green, Andrew J. Gentles, Weiguo Feng, Yue Xu, Chuong D. Hoang, Maximilian Diehn and Ash A. Alizadeh. Robust enumeration of cell subsets from tissue expression profiles.
}
\keyword{datasets}
