% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fardeep.R
\name{fardeep}
\alias{fardeep}
\title{Using the idea of least trimmed square to detect and remove outliers before estimating the coefficients.
A robust method for gene-expression deconvolution.}
\usage{
fardeep(X, Y, alpha1 = 0.1, alpha2 = 1.5, up = 10, low = 1, nn = TRUE,
  intercept = TRUE, lognorm = TRUE, permn = 100, QN = TRUE)
}
\arguments{
\item{alpha1}{parameter used to adjust the upper bound of outliers. Take value from 0 to 1, default 0.1.}

\item{alpha2}{parameter used to adjust the lower bound of outliers. Take value larger than 1, default 1.5.}

\item{up}{upper bound of parameter k in function alts, default 10.}

\item{low}{lower bound of parameter k in function alts, default 1.}

\item{nn}{whether coefficients are non-negative,default TRUE.}

\item{intercept}{whether intercept is included in model, default TRUE.}

\item{lognorm}{whether noise is log-normal distributed, default TRUE.}

\item{x}{input matrix of predictors with n rows and p columns.}

\item{y}{input vector of dependent variable.}
}
\value{
beta:  estimation of coefficients.

number_outlier:  number of outliers.

outlier_detect:  index of detected outliers.

X.new:  good observed points for independent variables.

Y.new:  good observed points for dependent variables.

k:  modified k (if the input value is not appropriate).
}
\description{
Using the idea of least trimmed square to detect and remove outliers before estimating the coefficients.
A robust method for gene-expression deconvolution.
}
\examples{
library(FARDEEP)
data(LM22)
data(mixture)
result = fardeep(LM22, mixture, alpha1 = 0.1, alpha2 = 1.5, up = 10, low = 1, nn = TRUE, intercept = TRUE, lognorm = TRUE, permn = 100, QN = TRUE)
coef = result$abs.beta
}
\references{
YuningHao, Ming Yan, Yu Lei and YuyingXie. Fast and Robust Deconvolution of Tumor Infiltrating Lymphocyte from Expression Profiles using Least Trimmed Squares.
}
\author{
YuningHao, Ming Yan, Yu Lei and YuyingXie
}
